refactoring the tests and the resources
